Crate interrupts

Basic interrupt handling structures and simple handler routines.

Constants

IRQ_BASE_OFFSET
The offset added to the first IRQ: 0x20.

Statics

APIC_TIMER_TICKS

IDT
The single system-wide Interrupt Descriptor Table (IDT).

Functions

deregister_interrupt
Returns an interrupt to the system by setting the handler to the default function. The application provides the current interrupt handler as a safety check. The function fails if the current handler and ‘func’ do not match

eoi
Send an end of interrupt signal, notifying the interrupt chip that the given interrupt request irq has been serviced.

init
Initializes the interrupt subsystem and sets up an initial Interrupt Descriptor Table (IDT).

init_ap
Similar to init(), but for APs to call after the BSP has already invoked init().

init_handlers_apic

init_handlers_pic

is_exception_handler_with_error_code
Returns true if the given address is the exception handler in the current IDT for any exception in which the CPU pushes an error code onto the stack.

register_interrupt
Registers an interrupt handler. The function fails if the interrupt number is already in use.

register_msi_interrupt
Returns an interrupt number assigned by the OS and sets its handler function. The function fails if there is no unused interrupt number.